+++ VRCVideoCacher/ConfigManager.cs
@@ -327,6 +327,7 @@ public class ConfigModel
     // Video Cacher
     public bool AutoUpdateVrcVideoCacher = true;
     public bool CloseToTray = true;
+    public bool StartMinimized = false;
     public bool CookieSetupCompleted = false;
 
     // Localization